- Fix for root converter attempting to delete internal (IMC) files directly. - Updated EQP flag name for a leg flag.

+++ xivModdingFramework/Mods/RootCloner.cs
@@ -150,7 +150,7 @@ namespace xivModdingFramework.Mods
                 var dPath = Destination.Info.GetRootFolder();
                 foreach (var mod in modlist.Mods)
                 {
-                    if (mod.fullPath.StartsWith(dPath))
+                    if (mod.fullPath.StartsWith(dPath) && !mod.IsInternal())
                     {
                         await _modding.DeleteMod(mod.fullPath, false);
                     }
